refactor: reorganise internal packages into modular domain structure
- Add internal/sensors/ with handlers, stores, SenML parser, InfluxDB writer, and models
- Add internal/platform/ with DB setup, migrations, device auth middleware, and health handler
- Remove old flat packages: auth, db, handler, influx, senml, store
- Update testutil to use platform package
- Update main.go to wire new package structure
- Update CLAUDE.md with package layout rules
