--- internal/handler/health.go
@@ -1,15 +1,0 @@
-package handler
-
-import (
-	"net/http"
-
-	"github.com/go-chi/render"
-)
-
-type HealthResponse struct {
-	Status string `json:"status"`
-}
-
-func Health(w http.ResponseWriter, r *http.Request) {
-	render.JSON(w, r, HealthResponse{Status: "ok"})
-}
