--- internal/handler/health_test.go
@@ -1,30 +1,0 @@
-package handler_test
-
-import (
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/fishhub-oss/fishhub-server/internal/handler"
-)
-
-func TestHealth(t *testing.T) {
-	req := httptest.NewRequest(http.MethodGet, "/health", nil)
-	w := httptest.NewRecorder()
-
-	handler.Health(w, req)
-
-	res := w.Result()
-	if res.StatusCode != http.StatusOK {
-		t.Fatalf("expected 200, got %d", res.StatusCode)
-	}
-
-	var body handler.HealthResponse
-	if err := json.NewDecoder(res.Body).Decode(&body); err != nil {
-		t.Fatalf("decode body: %v", err)
-	}
-	if body.Status != "ok" {
-		t.Fatalf("expected status ok, got %s", body.Status)
-	}
-}
