+++ internal/platform/db.go
@@ -1,4 +1,4 @@
-package db
+package platform
 
 import (
 	"database/sql"
