+++ internal/platform/middleware_test.go
@@ -1,4 +1,4 @@
-package auth_test
+package platform_test
 
 import (
 	"context"
@@ -6,24 +6,24 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/fishhub-oss/fishhub-server/internal/auth"
-	"github.com/fishhub-oss/fishhub-server/internal/store"
+	"github.com/fishhub-oss/fishhub-server/internal/platform"
+	"github.com/fishhub-oss/fishhub-server/internal/sensors"
 )
 
 type stubDeviceStore struct {
-	info store.DeviceInfo
+	info sensors.DeviceInfo
 	err  error
 }
 
-func (s *stubDeviceStore) LookupByToken(_ context.Context, _ string) (store.DeviceInfo, error) {
+func (s *stubDeviceStore) LookupByToken(_ context.Context, _ string) (sensors.DeviceInfo, error) {
 	return s.info, s.err
 }
 
-func TestAuthenticator(t *testing.T) {
-	validInfo := store.DeviceInfo{DeviceID: "device-uuid", UserID: "user-uuid"}
+func TestDeviceAuthenticator(t *testing.T) {
+	validInfo := sensors.DeviceInfo{DeviceID: "device-uuid", UserID: "user-uuid"}
 
 	next := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		info, ok := auth.DeviceFromContext(r.Context())
+		info, ok := sensors.DeviceFromContext(r.Context())
 		if !ok {
 			http.Error(w, "no device in context", http.StatusInternalServerError)
 			return
@@ -36,7 +36,7 @@ func TestAuthenticator(t *testing.T) {
 	})
 
 	t.Run("valid token passes through with device in context", func(t *testing.T) {
-		mw := auth.Authenticator(&stubDeviceStore{info: validInfo})
+		mw := platform.DeviceAuthenticator(&stubDeviceStore{info: validInfo})
 		req := httptest.NewRequest(http.MethodPost, "/readings", nil)
 		req.Header.Set("Authorization", "Bearer validtoken")
 		w := httptest.NewRecorder()
@@ -49,7 +49,7 @@ func TestAuthenticator(t *testing.T) {
 	})
 
 	t.Run("missing authorization header returns 401", func(t *testing.T) {
-		mw := auth.Authenticator(&stubDeviceStore{info: validInfo})
+		mw := platform.DeviceAuthenticator(&stubDeviceStore{info: validInfo})
 		req := httptest.NewRequest(http.MethodPost, "/readings", nil)
 		w := httptest.NewRecorder()
 
@@ -61,7 +61,7 @@ func TestAuthenticator(t *testing.T) {
 	})
 
 	t.Run("invalid token returns 401", func(t *testing.T) {
-		mw := auth.Authenticator(&stubDeviceStore{err: store.ErrTokenNotFound})
+		mw := platform.DeviceAuthenticator(&stubDeviceStore{err: sensors.ErrTokenNotFound})
 		req := httptest.NewRequest(http.MethodPost, "/readings", nil)
 		req.Header.Set("Authorization", "Bearer badtoken")
 		w := httptest.NewRecorder()
@@ -74,7 +74,7 @@ func TestAuthenticator(t *testing.T) {
 	})
 
 	t.Run("malformed authorization header returns 401", func(t *testing.T) {
-		mw := auth.Authenticator(&stubDeviceStore{info: validInfo})
+		mw := platform.DeviceAuthenticator(&stubDeviceStore{info: validInfo})
 		req := httptest.NewRequest(http.MethodPost, "/readings", nil)
 		req.Header.Set("Authorization", "notbearer token")
 		w := httptest.NewRecorder()
