+++ internal/sensors/handler.go
@@ -1,4 +1,4 @@
-package handler
+package sensors
 
 import (
 	"errors"
@@ -7,18 +7,41 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/fishhub-oss/fishhub-server/internal/auth"
-	"github.com/fishhub-oss/fishhub-server/internal/influx"
-	"github.com/fishhub-oss/fishhub-server/internal/senml"
 	"github.com/go-chi/render"
 )
 
+type TokenResponse struct {
+	Token    string `json:"token"`
+	DeviceID string `json:"device_id"`
+	UserID   string `json:"user_id"`
+}
+
+type TokensHandler struct {
+	Store  TokenStore
+	UserID string
+}
+
+func (h *TokensHandler) Create(w http.ResponseWriter, r *http.Request) {
+	result, err := h.Store.CreateToken(r.Context(), h.UserID)
+	if err != nil {
+		http.Error(w, "failed to create token", http.StatusInternalServerError)
+		return
+	}
+
+	render.Status(r, http.StatusCreated)
+	render.JSON(w, r, TokenResponse{
+		Token:    result.Token,
+		DeviceID: result.DeviceID,
+		UserID:   result.UserID,
+	})
+}
+
 type ReadingsHandler struct {
-	Writer influx.ReadingWriter
+	Writer ReadingWriter
 }
 
 func (h *ReadingsHandler) Create(w http.ResponseWriter, r *http.Request) {
-	device, ok := auth.DeviceFromContext(r.Context())
+	device, ok := DeviceFromContext(r.Context())
 	if !ok {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
 		return
@@ -30,11 +53,11 @@ func (h *ReadingsHandler) Create(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	reading, err := senml.Parse(body)
+	reading, err := ParseSenML(body)
 	if err != nil {
-		if errors.Is(err, senml.ErrEmptyPayload) ||
-			errors.Is(err, senml.ErrMissingBaseTime) ||
-			errors.Is(err, senml.ErrEmptyEntries) {
+		if errors.Is(err, ErrEmptyPayload) ||
+			errors.Is(err, ErrMissingBaseTime) ||
+			errors.Is(err, ErrEmptyEntries) {
 			http.Error(w, err.Error(), http.StatusBadRequest)
 			return
 		}
@@ -50,7 +73,7 @@ func (h *ReadingsHandler) Create(w http.ResponseWriter, r *http.Request) {
 		for _, m := range reading.Measurements {
 			fields[m.Name] = m.Value
 		}
-		if err := h.Writer.WriteReading(r.Context(), influx.Reading{
+		if err := h.Writer.WriteReading(r.Context(), Reading{
 			DeviceID:     device.DeviceID,
 			UserID:       device.UserID,
 			Timestamp:    time.Unix(reading.BaseTime, 0).UTC(),
